refactor(queue): extract AI-review orchestration into its own module
Part of #4013. Moves the AI-review lock cluster, the advisory-eligibility
and manifest/token resolution helpers, and the ~625-line
runAiReviewForAdvisory orchestration (grounding/RAG/enrichment/impact-map/
culture-profile/inline-findings) out of processors.ts into
queue/ai-review-orchestration.ts. processors.ts re-exports the symbols
still imported directly by existing tests.

Also fixes two grounding/enrichment wiring tests that registered a repo
outside the default GITTENSORY_REVIEW_REPOS allowlist, which silently
no-op'd their own "no installationId" assertions, and adds coverage for
the GitHub-token and PR-body optional-chain fallbacks.
